+++ app/src/main/java/org/sopt/at/presentaion/components/BottomNavBar.kt
@@ -8,21 +8,17 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 
-import androidx.compose.foundation.text.BasicText
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.sopt.at.R
-import org.sopt.at.data.remote.model.BottomNavItem
+import org.sopt.at.data.remote.BottomNavItem
 
 @Composable
 fun BottomNavBar(
